+++ environments/longhealth/longhealth.py
@@ -18,8 +18,8 @@ from typing import Any, Literal
 
 import verifiers as vf
 from datasets import Dataset
-from medarc_verifiers.rewards.mcq_accuracy import multiple_choice_accuracy
-from medarc_verifiers.utils.randomize_mcq import randomize_multiple_choice
+from medarc_verifiers.rewards.multiple_choice_accuracy import multiple_choice_accuracy
+from medarc_verifiers.utils.randomize_multiple_choice import randomize_multiple_choice
 from verifiers.utils.data_utils import BOXED_SYSTEM_PROMPT, THINK_BOXED_SYSTEM_PROMPT, extract_boxed_answer
 
 # Reuse the system prompt from the original LongHealth implementation
