+++ environments/medconceptsqa/medconceptsqa.py
@@ -3,8 +3,8 @@ from typing import Any
 import verifiers as vf
 from datasets import Dataset, load_dataset
 from medarc_verifiers.prompts import THINK_XML_SYSTEM_PROMPT, XML_SYSTEM_PROMPT, AnswerFormat
-from medarc_verifiers.rewards.mcq_accuracy import multiple_choice_accuracy
-from medarc_verifiers.utils.randomize_mcq import randomize_multiple_choice
+from medarc_verifiers.rewards.multiple_choice_accuracy import multiple_choice_accuracy
+from medarc_verifiers.utils.randomize_multiple_choice import randomize_multiple_choice
 from verifiers.utils.data_utils import BOXED_SYSTEM_PROMPT, THINK_BOXED_SYSTEM_PROMPT, extract_boxed_answer
 
 _VOCAB_CHOICES = ["atc", "icd10cm", "icd10proc", "icd9cm", "icd9proc"]
