+++ environments/medxpertqa/medxpertqa.py
@@ -1,8 +1,8 @@
 import verifiers as vf
 from datasets import load_dataset
 from medarc_verifiers.prompts import AnswerFormat
-from medarc_verifiers.rewards.mcq_accuracy import multiple_choice_accuracy
-from medarc_verifiers.utils.randomize_mcq import randomize_multiple_choice
+from medarc_verifiers.rewards.multiple_choice_accuracy import multiple_choice_accuracy
+from medarc_verifiers.utils.randomize_multiple_choice import randomize_multiple_choice
 from verifiers.utils.data_utils import extract_boxed_answer
