+++ environments/metamedqa/metamedqa.py
@@ -2,8 +2,8 @@ from typing import Dict
 
 import verifiers as vf
 from datasets import load_dataset
-from medarc_verifiers.rewards.mcq_accuracy import multiple_choice_accuracy
-from medarc_verifiers.utils.randomize_mcq import randomize_multiple_choice
+from medarc_verifiers.rewards.multiple_choice_accuracy import multiple_choice_accuracy
+from medarc_verifiers.utils.randomize_multiple_choice import randomize_multiple_choice
 
 
 def _build_prompt(question: str, options: Dict[str, str]) -> str:
