+++ medarc_verifiers/rewards/multiple_choice_accuracy.py
