+++ medarc_verifiers/utils/__init__.py
@@ -1,5 +1,9 @@
 from .download import download_file, medarc_cache_dir
-from .randomize_mcq import randomize_multiple_choice, randomize_multiple_choice_hf_map, randomize_multiple_choice_row
+from .randomize_multiple_choice import (
+    randomize_multiple_choice,
+    randomize_multiple_choice_hf_map,
+    randomize_multiple_choice_row,
+)
 
 __all__ = [
     "download_file",
