+++ tests/test_mcq_accuracy.py
@@ -1,6 +1,6 @@
 """Tests for the simplified MCQ accuracy grader."""
 
-from medarc_verifiers.rewards.mcq_accuracy import MCQAccuracyResult, multiple_choice_accuracy
+from medarc_verifiers.rewards.multiple_choice_accuracy import MCQAccuracyResult, multiple_choice_accuracy
 
 
 def test_anchored_final_answer_colon():
